Update card.js
added change in background color every time the page is refreshed.
added change in inner HTML to display the four suits of the poker deck. the inner HTML also changes color to 'red' when it's hearts and diamonds, and 'black' when it's clubs and spades.
added 'what to do' for each of the cards picked.
added visibility of the card when there are cards in the deck and when there is none.